+++ src/test/kotlin/me/bossm0n5t3r/leetcode/maximumSubarray/MaximumSubarrayTest.kt
@@ -1,10 +1,10 @@
-package maximumSubarray
+package me.bossm0n5t3r.leetcode.maximumSubarray
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class MaximumSubarrayTest {
-    private val maximumSubarray = MaximumSubarray()
+    private val maximumSubarray = MaximumSubarray.Solution()
 
     data class MaximumSubarrayTestData(
         val nums: IntArray,
